Add Botanical Encyclopedia app for browsing knowledge about plants (#96652)
## About The Pull Request

The app comes pre-installed on all botanist PDAs and can be manually
installed by anyone on any device.

I extracted the seed extractor UI into a seed table component that the
botanical encyclopedia now also uses.

I also made the seed extractor Scrap and Take buttons separate instead
of sharing a button with a toggle.

I'm not great at UI work so expect some rough edges and bugs. It does
work in testing.

Video of me testing the botanical encyclopedia:
https://streamable.com/um3yul

Video of me testing the seed extractor: https://streamable.com/5fn3ll
## Why It's Good For The Game

It allows players to access plant information without code-diving or
using the wiki.
## Changelog
:cl:
add: Added Botanical Encyclopedia app for browsing knowledge about
plants, which comes pre-installed on all botanist PDAs.
/:cl:
