+++ tgstation.dme
@@ -554,6 +554,7 @@
 #include "code\__HELPERS\atoms.dm"
 #include "code\__HELPERS\auxtools.dm"
 #include "code\__HELPERS\bitflag_lists.dm"
+#include "code\__HELPERS\botany.dm"
 #include "code\__HELPERS\chat.dm"
 #include "code\__HELPERS\chat_filter.dm"
 #include "code\__HELPERS\clients.dm"
@@ -3862,7 +3863,6 @@
 #include "code\modules\asset_cache\assets\research_designs.dm"
 #include "code\modules\asset_cache\assets\rtd.dm"
 #include "code\modules\asset_cache\assets\safe.dm"
-#include "code\modules\asset_cache\assets\seeds.dm"
 #include "code\modules\asset_cache\assets\sheetmaterials.dm"
 #include "code\modules\asset_cache\assets\slot_machines.dm"
 #include "code\modules\asset_cache\assets\supplypods.dm"
@@ -5974,6 +5974,7 @@
 #include "code\modules\modular_computers\file_system\programs\atmosscan.dm"
 #include "code\modules\modular_computers\file_system\programs\betting.dm"
 #include "code\modules\modular_computers\file_system\programs\borg_monitor.dm"
+#include "code\modules\modular_computers\file_system\programs\botanical_encyclopedia.dm"
 #include "code\modules\modular_computers\file_system\programs\bounty_board.dm"
 #include "code\modules\modular_computers\file_system\programs\budgetordering.dm"
 #include "code\modules\modular_computers\file_system\programs\card.dm"
